move RedemptionCode class to splash screen module
