+++ sphinx/screens/onboard/onboard-common/build.gradle
@@ -17,8 +17,6 @@ dependencies {
 
     implementation deps.javax.inject
 
-    implementation deps.jncryptor
-
     // needed to override moshi's use of 1.4.31
     implementation deps.kotlin.reflect
     implementation deps.square.moshi
